+++ frontend/src/metabase/lib/query/join.js
@@ -1,4 +1,4 @@
-import { add, update, remove, clear } from "./util";
+import { add, update, remove, clear } from "metabase-lib/lib/utils/query/util";
 
 // returns canonical list of Joins, with nulls removed
 export function getJoins(joins) {
